Fix the scope for loop variable (#385)
In 4.x the scope variable were no longer respected, so that a global loop variable couldn't be accessed from a function called within the loop. This commit restore the original behavior and makes sure that we also support local loop variable as well.
It is a breaking change because we are removing ScriptVariableLoop and ScriptVariableScope.Loop